// src/administrative-management/service-order/domain/enum/service-order-status.enum.ts
export enum ServiceOrderStatus {
    RECEBIDA = 'Recebida',
    EM_DIAGNOSTICO = 'Em diagnóstico',
    AGUARDANDO_APROVACAO = 'Aguardando aprovação',
    RECUSADA = 'Recusada',
    AGUARDANDO_INICIO = 'Aguardando início',
    EM_EXECUCAO = 'Em execução',
    FINALIZADA = 'Finalizada',
    ENTREGUE = 'Entregue',
  }